+++ changelog.md
@@ -20,6 +20,7 @@
     - Round word wrap widths to prevent floating point math errors
     - Fix word wrap calculations for long urls
     - Fix incorrect selection width for certain text style combinations
+  - Code editor: Fix goto line to correctly scroll to when clicking an error from output window
 - **Changed:**
   - `Character manager on row double click load in:` Now defaults to open in new tab
   - Display: Adjust word wrap break points to have better flow