+++ changelog.md
@@ -23,6 +23,7 @@
   - Code editor:
     - Fix goto line to correctly scroll to when clicking an error from output window
     - Fix tab strip to scroll when a tab is dragged over scroll buttons
+    - Fix goto/peek Definition menu and context items
   - Interface: Fix tab strip to scroll when a tab is dragged over scroll buttons
 - **Changed:**
   - `Character manager on row double click load in:` Now defaults to open in new tab